(maint) Update scooter pin to 4.0 and fix specs
Scooter was open sourced in 4.x. This commit updates the Gemfile, which
breaks a test relying on 3.x behavior. There is a fix for the spec that
mocks out scooter entirely, because initializing a scooter instance ends
up immediately trying to read ssl key files from the host object, and
dies when it can't construct them.