(PE-19049) Do not configure puppet agent service during upgrades.
Part of the upgrade process is validating that the agent services
upgrade properly when already running. This requires restarting puppet
agent prior to beginning the upgrading and checking its state after the
upgrade, so we need to not interfere with the service state in the
middle of the upgrade (whch is executed just like an install of a
particular version).