(PE-19049) Configure agents stopped during simple monolithic install
There is a new simple monolithic installation workflow, but now that
agent is managed, it also needs to configure the agent stopped. Commit
pulls out the code for this in a helper method and calls that in the two
spots simple_monolithic_install() and generic_install() where it is
needed.